refactor: gh-ost setting section
Close BYT-7362
