+++ frontend/src/components/Plan/components/Sidebar/GhostSection/FlagsForm/BoolFlag.vue
