+++ frontend/src/components/Plan/components/Sidebar/GhostSection/FlagsForm/FlagsForm.vue
