+++ frontend/src/components/Plan/components/Sidebar/GhostSection/index.ts
