+++ frontend/src/components/Plan/components/Sidebar/index.ts
@@ -1,6 +1,7 @@
 import PreBackupSection from "./PreBackupSection";
+import GhostSection from "./GhostSection";
 import Sidebar from "./Sidebar.vue";
 
 export default Sidebar;
 
-export { PreBackupSection };
+export { PreBackupSection, GhostSection };
