+++ frontend/src/locales/es-ES.json
@@ -278,7 +278,6 @@
       "ghost-parameters": "parámetros de gh-ost",
       "configure": "Configurar",
       "error": {
-        "x-status-task-is-not-editable": "La tarea {status} no es editable",
         "nothing-changed": "nada ha cambiado",
         "not-applicable": {
           "task-doesnt-meet-ghost-requirement": "La tarea no cumple con el requisito de gh-ost",
